完成系统日志

Closes #96

See merge request enterprise-data/404-not-found/enterprise-project-data-management-system!126

+++ backend/src/main/java/org/hfut/annotation/Log.java
@@ -13,5 +13,5 @@ public @interface Log {
 
     String serviceDescription();
 
-    int permission();
+    byte permission();
 }
